soil_moisture_monitor/requirements.txt
pyserial==2.6
slackclient==1.0.4
